# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервиса подачи заявок на 3D-печать МИЭМ НИУ ВШЭ.\\nАутентификация — OTP через email + JWT (Bearer token).",
        "title": "MIEM 3D Lab Print API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/admin/applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-applications"
                ],
                "summary": "Список всех заявок (admin)",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "enum": [
                                "new",
                                "in_review",
                                "printing",
                                "ready",
                                "issued",
                                "rejected",
                                "cancelled"
                            ],
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Фильтр по статусам",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск по имени / email",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID материала",
                        "name": "material_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата создания от (YYYY-MM-DD)",
                        "name": "created_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата создания до (YYYY-MM-DD)",
                        "name": "created_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Желаемая дата от (YYYY-MM-DD)",
                        "name": "desired_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Желаемая дата до (YYYY-MM-DD)",
                        "name": "desired_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Страница",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 20,
                        "description": "Записей на странице",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.AdminApplicationListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/applications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-applications"
                ],
                "summary": "Детали заявки (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.ApplicationDetailAdmin"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/applications/{id}/files/{file_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin-applications"
                ],
                "summary": "Скачать файл заявки (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID файла",
                        "name": "file_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Файл удалён",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/applications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Получить заявку",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.ApplicationDetailUser"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Требуется согласие",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.APIError"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.AdminApplicationListItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deadline_soon": {
                    "type": "boolean"
                },
                "desired_date": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "material_name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.AdminApplicationListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.AdminApplicationListItem"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.PaginationMeta"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer \u003caccess_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}